requirements.txt
# AstraFabric Platform Dependencies - Production Ready
# requirements.txt for Render.com deployment

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Limiter==3.5.0

# Production WSGI Server
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20
Flask-Migrate==4.0.5

# Security & Authentication
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.4
Flask-Talisman==1.1.0
passlib==1.7.4
argon2-cffi==23.1.0

# API & JSON
requests==2.31.0
marshmallow==3.20.1

# Background Tasks
celery==5.3.1
redis==4.6.0

# Monitoring & Logging
python-dotenv==1.0.0
structlog==23.1.0

# Testing & Quality
pytest==7.4.2
pytest-flask==1.2.0
safety==2.3.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3
